Added shifting bg.
The background can be chosen either as before, by defining a time frame or by following the shifting window. This works by choosing a value in minutes for bg_distance_from_window which is the distance of the bg start from the window. This works with a fixed or shifting window. I lest the old option in case there is need for a specific type of  bg.